+): 将单元测试代码移到unit_test目录下 +): Fix 编译警告(关于char*和const char*转换, signed和unsigned integer比较)

+++ leetcode/longest_substring_without_repeating_char.cpp
@@ -12,12 +12,12 @@ int lengthOfLongestSubstring(string s)
 
     char map_char[256] = {0};
 
-    for(int i=0; i<s.length(); i++)
+    for(size_t i=0; i<s.length(); i++)
     {
         len = 0;
         memset(map_char, 0, 256);
 
-        for(int j=i; j<s.length(); j++)
+        for(size_t j=i; j<s.length(); j++)
         {
             if(map_char[s[j]] == 0)
             {
